% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decode-base.R, R/decode-generic.R
\name{decode.spress_index}
\alias{decode.spress_index}
\alias{decode.data.frame}
\alias{decode.double}
\alias{decode.integer}
\alias{decode}
\title{Decode a Hilbert curve index to coordinates}
\usage{
\method{decode}{spress_index}(x, ...)

\method{decode}{data.frame}(x, ..., extent, index = length(x), n = 4L)

\method{decode}{double}(x, ..., extent, n = 4L)

\method{decode}{integer}(x, ..., extent, n = 4L)

decode(x, ..., extent, n = 4L)
}
\arguments{
\item{x}{A \code{data.frame} or \code{numeric} vector}

\item{...}{Unused}

\item{extent}{Extent of the original dataset.
If encoded with \code{encode}, then
the attribute \code{spress_extent} is used.}

\item{index}{Index or name of column containing the Hilbert curve index}

\item{n}{Level of space division, must be between 0 and 16.
If encoded with \code{encode}, then
the attribute \code{spress_n} is used.}
}
\description{
Decode a Hilbert curve index to coordinates
}
