% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{.pull_crs}
\alias{.pull_crs}
\title{Pull CRS out of \code{sf} object.}
\usage{
.pull_crs(sf_obj)
}
\arguments{
\item{sf_obj}{\code{sf}/\code{sfc}/\code{sfg} object}
}
\value{
CRS object
}
\description{
Pull CRS out of \code{sf} object.
}
\keyword{internal}
