% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{.pull_geometry}
\alias{.pull_geometry}
\title{Pull geometry out of \code{sf} object}
\usage{
.pull_geometry(sf_obj)
}
\arguments{
\item{sf_obj}{\code{sf}/\code{sfc}/\code{sfg} object}
}
\value{
Geometry object
}
\description{
Pull geometry out of \code{sf} object
}
\keyword{internal}
