% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-sf.R
\name{.pull_geotype}
\alias{.pull_geotype}
\title{Pull geometry type out of \code{sf} object.}
\usage{
.pull_geotype(sf_obj)
}
\arguments{
\item{sf_obj}{\code{sf}/\code{sfc}/\code{sfg} object}
}
\value{
Geometry type string
}
\description{
Pull geometry type out of \code{sf} object.
}
\keyword{internal}
