% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode-base.R, R/encode-generic.R
\name{encode.data.frame}
\alias{encode.data.frame}
\alias{encode.numeric}
\alias{encode}
\title{Encode coordinates to a Hilbert curve index}
\usage{
\method{encode}{data.frame}(x, ..., n = 4L, coords = c(1, 2), attach = FALSE)

\method{encode}{numeric}(x, y, ..., n = 4L)

encode(x, ..., n = 4L)
}
\arguments{
\item{x}{A \code{data.frame} or \code{numeric} vector}

\item{...}{Unused}

\item{n}{Level of space division, must be between 0 and 16.}

\item{coords}{Column indices or names corresponding to coordinates}

\item{attach}{If \code{TRUE}, attaches index to data frame as new column}

\item{y}{A \code{numeric} vector}
}
\description{
Encode coordinates to a Hilbert curve index
}
